changes type double to float; rename some function parametors and variables for norminette

+++ inc/light.h
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   light.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cle-tron <cle-tron@student.42barcelon      +#+  +:+       +#+        */
+/*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/18 18:39:49 by cle-tron          #+#    #+#             */
-/*   Updated: 2024/12/18 19:06:24 by cle-tron         ###   ########.fr       */
+/*   Updated: 2024/12/19 17:30:35 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIGHT_H
 # define LIGHT_H
 
-int		color_converter(int color[3], double intensity);
+int		color_converter(int color[3], float intensity);
 void	lightning(t_hit_rec *rec, t_data *data);
 
 #endif
