+++ src/ray/generate_ray.c
@@ -6,29 +6,29 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 20:14:21 by wlin              #+#    #+#             */
-/*   Updated: 2024/11/16 20:29:02 by wlin             ###   ########.fr       */
+/*   Updated: 2024/12/19 17:30:35 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
-double	screen_x_to_viewport(int x, double aspect_ratio, double fov_scale)
+float	screen_x_to_viewport(int x, float aspect_ratio, float fov_scale)
 {
-	return ((2.0 * (x + 0.5) / (double)WIDTH - 1.0) * aspect_ratio * fov_scale);
+	return ((2.0 * (x + 0.5) / (float)WIDTH - 1.0) * aspect_ratio * fov_scale);
 }
 
-double	screen_y_to_viewport(int y, double fov_scale)
+float	screen_y_to_viewport(int y, float fov_scale)
 {
-	return ((1.0 - 2.0 * (y + 0.5) / (double)HEIGHT) * fov_scale);
+	return ((1.0 - 2.0 * (y + 0.5) / (float)HEIGHT) * fov_scale);
 }
 
 t_ray generate_ray(t_cam camera, t_view view_params, int x, int y)
 {
 	t_ray	ray;
-	double	view_right[3];
-	double	view_up[3];
-	double	normalized_x;
-	double	normalized_y;
+	float	view_right[3];
+	float	view_up[3];
+	float	normalized_x;
+	float	normalized_y;
 	
 	// Ray starts from camera
 	vec_copy(ray.origin, camera.xyz);
